# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech-Admin-Api",
        "contact": {
            "name": "Tech-Admin",
            "url": "  http://tech-admin.com",
            "email": "  @gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/acesso": {
            "get": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Lista todos os acessos",
                "description": "Retorna todos os acessos",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Retorna todos os acessos"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Cria um novo acesso",
                "description": "Cria um novo acesso",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do acesso",
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "description": "Descrição do acesso",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status do acesso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o acesso criado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            }
        },
        "/api/acesso/{id}": {
            "get": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Mostra um acesso",
                "description": "Mostra um acesso",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Atualiza um acesso",
                "description": "Atualiza um acesso",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do acesso",
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "description": "Descrição do acesso",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status do acesso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o acesso atualizado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Deleta um acesso",
                "description": "Deleta um acesso",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso deletado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            }
        },
        "/api/categoria": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Lista todas as categorias",
                "description": "Retorna todas as categorias",
                "operationId": "indexCategoria",
                "responses": {
                    "200": {
                        "description": "Retorna todas as categorias"
                    },
                    "default": {
                        "description": "Caso não encontre a categoria"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Cria uma nova categoria",
                "description": "Cria uma nova categoria",
                "operationId": "storeCategoria",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "description": "Nome da categoria",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna a categoria criada"
                    },
                    "default": {
                        "description": "Caso não encontre a categoria"
                    }
                }
            }
        },
        "/api/categoria/{id}": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Lista uma categoria",
                "description": "Retorna uma categoria",
                "operationId": "showCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a categoria"
                    },
                    "default": {
                        "description": "Caso não encontre a categoria"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Atualiza uma categoria",
                "description": "Atualiza uma categoria",
                "operationId": "updateCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "description": "Nome da categoria",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna a categoria atualizada"
                    },
                    "default": {
                        "description": "Caso não encontre a categoria"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Deleta uma categoria",
                "description": "Deleta uma categoria",
                "operationId": "destroyCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a categoria deletada"
                    },
                    "default": {
                        "description": "Caso não encontre a categoria"
                    }
                }
            }
        },
...
    }
}